fixed problems with multigrid hierarchies and refinement, which i unfortunately introduced by changing the behaviour of autogenerated edges and faces in grid_connection_managment. Those changes are now reverted.

+++ ugbase/lib_grid/multi_grid.h
@@ -229,7 +229,7 @@ class MultiGrid : public Grid, public GridObserver
 		inline uint num_levels() const	{return m_hierarchy.num_subsets();}
 
 		template <class TElem> inline
-		uint num(int level) const		{return m_hierarchy.num<TElem>(level);}
+		size_t num(int level) const		{return m_hierarchy.num<TElem>(level);}
 
 		template <class TElem> inline
 		typename geometry_traits<TElem>::iterator
